added test for resampling floats
fixed bug when resampling floats

git-svn-id: file:///mnt/svn/libmusicaccess/trunk@1094 0c85f470-1217-11de-aba5-9925624bc3b7

+++ src/musicaccess/resample.cpp
@@ -451,7 +451,7 @@ namespace musicaccess
         float* samples = *samplePtr;
         for (unsigned int i = 0; i < frameCount; i++)
         {
-            int32_t tmpVal = 0;
+            float tmpVal = 0;
             int offset = channelCount * i;
             for (unsigned int j = 0; j < channelCount; j++)
             {
